Fixed the missing problem where specular maps would not be used: The maps were set to the shader in the material with setFloat() which should have been setInt()

+++ OpenGL_Two/OpenGL_Two/Source/Materials/Material_A.h
@@ -16,6 +16,7 @@ public:
 		glEnable(GL_DEPTH_TEST);
 	}
 
+
 	void InitDraw(glm::mat4 view, glm::mat4 projection, glm::mat4 model, Camera* drawCam, std::vector<Light*> lights) override {
 		shaderOne.use();
 
@@ -38,7 +39,8 @@ public:
 			else if (name == "texture_normal") number = std::to_string(normalNr++);
 			else if (name == "texture_height") number = std::to_string(heightNr++);
 
-			shaderOne.setFloat((name + number).c_str(), i);
+			//This NEEDS to be setInt NOT setFloat!!!
+			shaderOne.setInt((name + number).c_str(), i);
 			glBindTexture(GL_TEXTURE_2D, textures[i].id);
 		}
